+++ packages/antd/modules/index.d.ts
@@ -1,8 +1,7 @@
 import {BasicConfig} from "@react-awesome-query-builder/ui";
-import {type AntdWidgets} from "./widgets";
+export {AntdWidgets} from "./widgets";
 
 export declare const AntdConfig: BasicConfig;
-export declare const AntdWidgets: AntdWidgets;
 
 // re-export
 export * from "@react-awesome-query-builder/ui";