+++ packages/bootstrap/.babelrc.js
@@ -9,10 +9,10 @@ const config = {
     "@babel/preset-react"
   ],
   "plugins": [
-    ["@babel/plugin-proposal-class-properties", { "loose": true }],
-    ["@babel/plugin-proposal-private-methods", { "loose": true }],
+    ["@babel/plugin-transform-class-properties", { "loose": true }],
+    ["@babel/plugin-transform-private-methods", { "loose": true }],
     ["@babel/plugin-transform-runtime", { "loose": true }],
-    ["@babel/plugin-proposal-private-property-in-object", { "loose": true }]
+    ["@babel/plugin-transform-private-property-in-object", { "loose": true }]
   ],
   "env": {
     "production": {