+++ packages/bootstrap/modules/index.d.ts
@@ -1,8 +1,7 @@
 import {BasicConfig} from "@react-awesome-query-builder/ui";
-import {type BootstrapWidgets} from "./widgets";
+export {BootstrapWidgets} from "./widgets";
 
 export declare const BootstrapConfig: BasicConfig;
-export declare const BootstrapWidgets: BootstrapWidgets;
 
 // re-export
 export * from "@react-awesome-query-builder/ui";