+++ packages/material/modules/index.d.ts
@@ -1,8 +1,7 @@
 import {BasicConfig} from "@react-awesome-query-builder/ui";
-import {type MaterialWidgets} from "./widgets";
+export {MaterialWidgets} from "./widgets";
 
 export declare const MaterialConfig: BasicConfig;
-export declare const MaterialWidgets: MaterialWidgets;
 
 // re-export
 export * from "@react-awesome-query-builder/ui";