+++ packages/sandbox/vite.config.ts
@@ -13,6 +13,16 @@ export default defineConfig({
   ],
   build: {
     sourcemap: true,
+    rollupOptions: {
+      // https://github.com/vitejs/vite/issues/15012
+      onwarn(warning, defaultHandler) {
+        if (warning.code === 'SOURCEMAP_ERROR') {
+          return
+        }
+
+        defaultHandler(warning)
+      },
+    },
     minify: false,
   },
 });