+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
+- 6.5.2
+  - Updated dependencies. `@babel/runtime` is now dep for core package (PR #1051) (issue #964)
 - 6.5.1
   - Fix missing config file in antd moment date widgets (PR #1046) (issue #1048)
   - Fix SpEL import not_between (PR #1050) (issue #1047)
 - 6.5.0
   - Breaking: `Utils.validateTree` returns array of errors (PR #1034)
+  - Breaking: `Utils.sanitizeTree` returns object `{fixedTree, fixedErrors}` (PR #1034)
   - Breaking: `removeIncompleteRulesOnLoad` by default is false
   - Added config `removeEmptyRulesOnLoad` (default true) (PR #1034)
   - Deprecated `Utils.validateAndFixTree` and `Utils.checkTree` (and removed type defs).